+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * [#627](https://github.com/bbatsov/rubocop/pull/627): Fix counting of slashes in complicated regexps in `RegexpLiteral` cop. ([@jonas054][])
 * [#638](https://github.com/bbatsov/rubocop/issues/638): Fix bug in auto-correct that changes `each{ |x|` to `each d o |x|`. ([@jonas054][])
 * [#418](https://github.com/bbatsov/rubocop/issues/418): Stop searching for configuration files above the work directory of the isolated environment when running specs. ([@jonas054][])
+* Fix error on implicit match conditionals (e.g. `if /pattern/; end`) in `MultilineIfThen`. ([@agrimm][])
 
 ## 0.15.0 (06/11/2013)
 
@@ -587,3 +588,4 @@
 [@fancyremarker]: https://github.com/fancyremarker
 [@sinisterchipmunk]: https://github.com/sinisterchipmunk
 [@vonTronje]: https://github.com/vonTronje
+[@agrimm]: https://github.com/agrimm